Move InstallPackage from cmd/packagecmd to cmd/packages
